Inline StratifierManager get
