+++ libhist/CMakeLists.txt
@@ -1,12 +1,12 @@
-add_library(libhist INTERFACE)
+add_library(libhist VectorStratifier.cpp)
 
 target_include_directories(libhist
-  INTERFACE
+  PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
-    $<INSTALL_INTERFACE:include>   
+    $<INSTALL_INTERFACE:include>
 )
 
 target_link_libraries(libhist
-  INTERFACE
+  PUBLIC
     libutils
 )
