+++ libhist/StratifierManager.h
@@ -8,9 +8,7 @@
 #include "AnalysisLogger.h"
 #include "IHistogramStratifier.h"
 #include "KeyTypes.h"
-#include "ScalarStratifier.h"
 #include "StratifierRegistry.h"
-#include "VectorStratifier.h"
 
 namespace analysis {
 
@@ -18,36 +16,13 @@ class StratifierManager {
   public:
     explicit StratifierManager(StratifierRegistry &registry) : registry_(registry) {}
 
-    IHistogramStratifier &get(const StratifierKey &key) {
-        log::debug("StratifierManager::get", "Attempting to get stratifier for key:", key.str());
-        auto it = cache_.find(key);
-        if (it == cache_.end()) {
-            log::info("StratifierManager::get", "Creating new stratifier for key:", key.str());
-
-            std::unique_ptr<IHistogramStratifier> stratifier;
-            StratifierType type = registry_.findSchemeType(key);
-
-            if (type == StratifierType::kScalar) {
-                stratifier = std::make_unique<ScalarStratifier>(key, registry_);
-            } else if (type == StratifierType::kVector) {
-                stratifier = std::make_unique<VectorStratifier>(key, registry_);
-            } else {
-                log::fatal("StratifierManager::get", "Unknown or unregistered stratifier configuration:", key.str());
-            }
-
-            it = cache_.emplace(key, std::move(stratifier)).first;
-            log::debug("StratifierManager::get", "Successfully created and cached stratifier for key:", key.str());
-        } else {
-            log::debug("StratifierManager::get", "Found cached stratifier for key:", key.str());
-        }
-        return *it->second;
-    }
+    IHistogramStratifier &get(const StratifierKey &key);
 
   private:
     StratifierRegistry &registry_;
     std::unordered_map<StratifierKey, std::unique_ptr<IHistogramStratifier>> cache_;
 };
 
-} // namespace analysis
+}
 
 #endif
